fix(depthwise): compute per-channel effective scales in double
The weight scales are float32, so the product stayed float32 and the
emitted Q31 multipliers differed from TFLite's. For
depthwise_conv_dilated_1d_k7_d8_c24_s16 channel 3 lands on the Q15
reduction tie (0x77FF8000 -> 0x7800 instead of TFLite's 0x77FF), which
put 78 outputs 1 LSB off the golden.

Refs #213

+++ helia_core_tester/generation/ops/ConvolutionFunctions/depthwise_conv.py
@@ -809,7 +809,9 @@ class OpDepthwiseConv(OperationBase):
         # Calculate effective scales: (input_scale * weight_scale) / output_scale
         if per_channel and isinstance(weight_scale, np.ndarray):
             # Per-channel: effective_scale[i] = (input_scale * weight_scale[i]) / output_scale
-            effective_scales = (input_scale * weight_scale) / output_scale
+            # in double precision, as TFLite computes it. The weight scales are float32, and
+            # float32 arithmetic can move a Q31 multiplier onto a different Q15 rounding for s16.
+            effective_scales = (input_scale * weight_scale.astype(np.float64)) / output_scale
             effective_quant = {
                 'scale': effective_scales,
                 'zero_point': output_quant.get('zero_point', 0),
